Actual output should not be ignored with sasldblistusers2
Not sure if this is different for other platforms but on EL 6 the
current code would ignore lines with actual output which obviously
isn't very desirable. Not sure if other platforms show some kind
of headers / footers with meta-information that need be stripped
away. But then again, unless that meta stuff is in the format of
.*@.*:.*, the match will be bad anyway and therefore not change the
behavior at all.